Enforce magic check before using any particle group
Particle group IDs now pack the index + a magic number into one int32_t.

This fixes fire particles not appearing when lv5 beehive is burning at high framerates (>100 fps).

May fix other particle effects with timed emitters if the game merely checked that the group wasn't -1 without looking at the magic number.

+++ src/Headers/structs.h
@@ -298,7 +298,7 @@ struct ObjNode
 	short				SplineObjectIndex;		// index into gSplineObjectList of this ObjNode
 
 	short				EffectChannel;			// effect sound channel index (-1 = none)
-	short				ParticleGroup;
+	int32_t				ParticleGroup;
 };
 typedef struct ObjNode ObjNode;
